+++ constants.js
@@ -1,0 +1,28 @@
+module.exports = {
+  API_URL: 'https://secure.splitwise.com/api/v3.0/',
+  PROP_NAMES: {
+    CURRENCIES: 'currencies',
+    CATEGORIES: 'categories',
+    USER: 'user',
+    GROUPS: 'groups',
+    GROUP: 'group',
+    EXPENSES: 'expenses',
+    EXPENSE: 'expense',
+    FRIENDS: 'friends',
+    FRIEND: 'friend',
+    NOTIFICATIONS: 'notifications',
+    SUCCESS: 'success'
+  },
+  ID_PARAM_NAMES: {
+    USER: 'userID',
+    GROUP: 'groupID',
+    EXPENSE: 'expenseID',
+    FRIEND: 'friendID'
+  },
+  METHOD_VERBS: {
+    GET: 'GET',
+    POST: 'POST',
+    PUT: 'PUT',
+    DELETE: 'DELETE'
+  }
+}
